Updated MessageTrait to add setters
Protocol, headers, and body are all now mutable by default; brought in
methods from WriteableMessageTrait to complete the trait.

Also, renamed `getHeaderAsArray()` to `getHeaderLines()` per the change
in the interface.
